Added security for rest api endpoints

+++ src/main/java/com/moger/crudproject/controller/StudentController.java
@@ -1,11 +1,14 @@
 package com.moger.crudproject.controller;
 
-import com.moger.crudproject.DAO.StudentDAO;
+import com.moger.crudproject.dao.StudentDAO;
 import com.moger.crudproject.entity.Student;
+import com.moger.crudproject.exception.DataNotFoundException;
+import com.moger.crudproject.exception.StudentNotFoundException;
 import jakarta.validation.Valid;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
+import org.springframework.http.HttpStatus;
 import java.util.List;
 
 @RestController
@@ -18,36 +21,54 @@ public class StudentController {
 
     @PostMapping(value = "/students", consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public String createStudent(@RequestBody @Valid Student student) {
+    public ResponseEntity<String> createStudent(@RequestBody @Valid Student student) {
+
         Long studentId = studentDAO.save(student);
-        return String.format("Created student with id %d", studentId);
+        return new ResponseEntity<>(String.format("Created student with id %d", studentId), HttpStatus.CREATED);
     }
 
     @GetMapping(value = "/students/{id}")
-    public Student findStudent(@PathVariable Long id) {
-        return studentDAO.findStudentById(id);
+    public ResponseEntity<Student> findStudent(@PathVariable Long id) {
+
+        List<Student> students = studentDAO.getAllStudents().stream().filter(n -> n.getId()==id).toList();
+        if (students.isEmpty())
+            throw new StudentNotFoundException(String.format("Student with %d does not exist!", id));
+        return new ResponseEntity<>(studentDAO.findStudentById(id),HttpStatus.OK);
     }
 
     @GetMapping(value = "/students")
     public List<Student> findAll() {
+
         return studentDAO.getAllStudents();
     }
 
     @PutMapping(value = "/students/{id}")
-    public String updateStudent(@PathVariable Long id, @RequestBody @Valid Student student) {
-        Student beforeUpdateStudent = studentDAO.updateStudent(id, student);
-        student.setId((int) beforeUpdateStudent.getId());
-        return beforeUpdateStudent + String.format(" now changed to " + student);
+    public ResponseEntity<Object> updateStudent(@PathVariable Long id, @RequestBody @Valid Student student) {
+
+        List<Student> students = studentDAO.getAllStudents().stream().filter(n -> n.getId()==id).toList();
+        if (students.isEmpty())
+            throw new DataNotFoundException(String.format("Student with %d does not exist!", id));
+
+        Student beforeUpdateStudent = studentDAO.findStudentById(id);
+        student.setId(id);
+        return new ResponseEntity<>(beforeUpdateStudent + String.format(" now changed to " + student),HttpStatus.OK);
     }
 
     @DeleteMapping(value = "/students/{id}")
-    public String deleteStudent(@PathVariable Long id) {
+    public ResponseEntity<Object> deleteStudent(@PathVariable Long id) {
+
+        List<Student> students = studentDAO.getAllStudents().stream().filter(n -> n.getId() == id).toList();
+        if (students.isEmpty())
+            throw new StudentNotFoundException(String.format("Student with %d does not exist!", id));
         studentDAO.deleteStudent(id);
-        return String.format("Student with id %d deleted",id);
+
+        String message = String.format("Student with %d deleted", id);
+        return new ResponseEntity<>(message, HttpStatus.OK);
     }
 
     @DeleteMapping(value = "/students")
     public String deleteAll() {
+
         int rowsAffected = studentDAO.deleteAllStudents();
         return String.format("%d records removed", rowsAffected);
     }
